fix: do not stringify headers for logging
The frame headers are passed through the `stringifyHeader` for trace
logging if a logger exists.  Unfortunately a logger is set by default
so this always occurs.

If we need to log these things as a string, checking that the logger
is enabled is one option:

```js
if (this.log.enabled) {
  // do expensive logging operation
}
```

Another is adding a custom formatter to the log class, or just having
the header object appear in the console.