fix: remove abortable iterator from muxer
We use an abortable source for every connection - this introduces
a lot of latency.

This PR removes abortable iterator to just listen on the abort event
on the shutdown controller's abort signal and close the connection's
iterator.

This is the same thing that abortable source does except we don't
race every `next` from the source.

I've added a simple benchmark - in my testing it increases
throughput by 30% or so, this seems to translate to 10-15% in "real world"
performance using the `@libp2p/perf` testing protocol.