fix: establish RTT at start of connection
A connection ping is required to establish the RTT - as implemented
the ping only happens during the keep alive loop that doesn't kick
in until after 30s by default.

The change is to ping immediately on sinking a source - this means
we are more likely to increase the send window on the remote which
increases the amount of data they send and as such, throughput.