fix: send data during graceful close.
When a stream is closed gracefully, it's status goes from `'open'`
to `'closing'` then to either `'closed'`, `'aborted'` or `'reset'`.

While it's `'closing'` we should still try to send any queued data,
this can be aborted by calling `.abort` on the stream or by the
signal passed to `.close` firing the `'abort'` event.

This change makes the tests added in https://github.com/libp2p/js-libp2p/pull/2398
pass.